+++ src/asdf_standard/_versioning.py
@@ -1,9 +1,9 @@
 import os
 import warnings
 
-from .exceptions import DevCoreSchemasWarning
+from .exceptions import UnstableCoreSchemasWarning
 
-DEV_ENV_VAR = "ASDF_DEV_CORE_SCHEMAS"
+UNSTABLE_ENV_VAR = "ASDF_UNSTABLE_CORE_SCHEMAS"
 
 stable_core_schema_versions = (
     "1.0.0",
@@ -15,28 +15,30 @@ stable_core_schema_versions = (
     "1.6.0",
 )
 
-dev_core_schema_versions = ("1.7.0",)
+unstable_core_schema_versions = ("1.7.0",)
 
 
-_dev_core_schema_warning = (
-    f"{DEV_ENV_VAR} is enabled and development versions of core schemas "
-    f"{dev_core_schema_versions} are in use. Files produced with development "
+_unstable_core_schema_warning = (
+    f"{UNSTABLE_ENV_VAR} is enabled and unstable versions of core schemas "
+    f"{unstable_core_schema_versions} are in use. Files produced with unstable/development "
     "schemas may later be made invalid due to ongoing schema changes."
 )
 
 
 def get_supported_core_schema_versions():
-    if get_dev_supported():
-        warnings.warn(_dev_core_schema_warning, DevCoreSchemasWarning)
-        return stable_core_schema_versions + dev_core_schema_versions
+    if get_unstable_supported():
+        return stable_core_schema_versions + unstable_core_schema_versions
     return stable_core_schema_versions
 
 
-def get_dev_supported():
-    value = os.getenv(DEV_ENV_VAR)
+def get_unstable_supported():
+    value = os.getenv(UNSTABLE_ENV_VAR)
     if not value:
         return False
     try:
-        return bool(int(value))
+        unstable_supported = bool(int(value))
     except ValueError:
-        raise ValueError(f"{DEV_ENV_VAR} value {value} is not supported, use either 0 or 1")
+        raise ValueError(f"{UNSTABLE_ENV_VAR} value {value} is not supported, use either 0 or 1")
+    if unstable_supported:
+        warnings.warn(_unstable_core_schema_warning, UnstableCoreSchemasWarning)
+    return unstable_supported