+++ CHANGES.rst
@@ -5,9 +5,9 @@
   These are only used by legacy schemas and are not to be
   used for new schemas. See asdf-transform-schemas for
   newer versions of these schemas [#485]
-- Support setting ASDF_DEV_CORE_SCHEMAS environment variable
-  to enable support for upcoming (development) versions of
-  core schemas. Be warned that writing files with development
+- Support setting ASDF_UNSTABLE_CORE_SCHEMAS environment variable
+  to enable support for upcoming (currently unstable) versions of
+  core schemas. Be warned that writing files with unstable/development
   versions of core schemas is not recommended as these schemas
   may change during development [#479]